API: minReadyInstanceCount must be int32
Make type of instance count consistent in all places.
Add into pods manager interface methods for controlling instance count.

And remove useless copy-paste tests for remote nodes which tests nothing.
Remote reconcilers are broken too, proper e2e will be added as part of fix.

+++ test/r8r/components_test.go
@@ -281,7 +281,7 @@ var _ = Describe("Components reconciler", Label("reconciler"), func() {
 
 		By("Creating minimal Ytsaurus spec", func() {
 			ytBuilder = &testutil.YtsaurusBuilder{
-				MinReadyInstanceCount: ptr.To(0), // Do not wait any pods.
+				MinReadyInstanceCount: ptr.To(int32(0)), // Do not wait any pods.
 				Images:                testutil.TestImages,
 				Namespace:             namespace,
 			}
